Primitive attempt to get replies via search
